pyproject.toml
[tool.poetry]
name = "edcornard-chatbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = ">=0.109.1"
uvicorn = { version = "^0.23.2", extras = [ "standard" ] }
python-multipart = "^0.0.9"
llama-index-vector-stores-chroma = "^0.1.8"
llama-index-embeddings-gemini = "^0.1.6"
llama-index-llms-gemini = "^0.1.10"
llama-index-llms-ollama = "^0.1.3"
llama-index-vector-stores-qdrant = "^0.2.8"
ollama = "^0.2.0"
llama-index-llms-azure-openai = "^0.1.8"
llama-index-embeddings-azure-openai = "^0.1.9"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
opentelemetry-api = "^1.25.0"
opentelemetry-sdk = "^1.25.0"
aioredis = "^2.0.1"
redis = "^5.0.5"
sqlalchemy = "^2.0.30"
opentelemetry-exporter-otlp = "^1.25.0"
opentelemetry-exporter-otlp-proto-grpc = "^1.25.0"
opentelemetry-instrumentation = "^0.46b0"
opentelemetry-instrumentation-logging = "^0.46b0"
pika = "^1.3.2"
git-filter-repo = "^2.38.0"
python-jose = "^3.3.0"
httpx = "^0.27.0"
pydantic = {extras = ["email"], version = "^2.7.4"}
jinja2 = "^3.1.4"
resend = "^2.1.0"
pyjwt = "^2.8.0"
python-dotenv = "^1.0.1"
numpy = "<2.0"
boto3 = "^1.34.136"
aiostream = "^0.6.1"
beautifulsoup4 = "^4.12.3"
pandas = "^2.2.2"
reportlab = "^4.2.2"
html2text = "^2024.2.26"
llama-index-readers-file = "^0.1.27"
ftfy = "^6.2.0"
fpdf2 = "^2.7.9"
langchain = "^0.2.10"
langchain-community = "^0.2.9"
langchain-core = "^0.2.22"
llama-index-core = "0.10.37.post1"
llama-index-postprocessor-cohere-rerank = "0.1.6"
pydantic-settings = "^2.4.0"
llama-index-llms-anthropic = "0.1.15"
llama-index-embeddings-huggingface = "0.2.3"


[tool.poetry.group.dev.dependencies]
black = "^24.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
